Add files via upload
This is a Dr.Racket (Scheme) code, which finds out day of the date, using the Zeller's Rule to find the day of the week. It can accept wrong dates, such as 31st november, but will tell the user about it being invalid. This program can stretch to any date from 1752 October to 9999 December 31st. 
Please feel free to let me know if there are any methods I can increase the efficiency of the code or there is any bug.

Thank you ^.^